backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
pymupdf==1.23.8
python-docx==0.8.11
openpyxl==3.10.10
pandas==2.1.3
langgraph==0.0.10
langchain==0.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
cors-headers==4.3.1
pytest==8.3.0
pytest-asyncio==0.23.5
httpx==0.27.0
